Extend caching mechanism to support plugin.
Cache managers need the cache folder to be set.
In the CLI app we set the cache folder to the output file as it was
previously.

Caching functions also requires file name of caching hing, which is used
when saving the cache file itself. BufferInputData didn't have this
hing, so we had to add it. Cache file hint is requested by the function
call.