Pass InputData object to stream decompression.
Stream compression is in its nature compressing different data every time. That's why we shouldn't work with cached instance options which includes InputData object.

This was a problem when multiple threads tried to compress different data, but were overriding each other InputData info.

Because of this we changed compressVoxels() signature to access InputData.